学习路径助手/requirements.txt
python>=3.11
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.28.2
pyodbc>=4.0.39          # SQL Server驱动
openai>=1.3.7
python-dotenv>=1.0.0
pandas>=2.1.4
requests>=2.31.0